added decks more refactoring
