+++ src/card/mod.rs
@@ -1,2 +1,1 @@
 pub mod ranks;
-
