+++ src/card/ranks/mod.rs
@@ -1,6 +1,5 @@
 mod euchre_rank;
 mod standard52_rank;
-mod generic;
 
 use crate::fluent::Named;
 use std::str::FromStr;
